Don't wrap 2 byte unsigned values in Java shorts Indicate in history broadcast whether the broadcasted item has been resynced Close confirmation dialogs on disconnect Made service more resistant against exceptions in remote process

+++ app/src/main/java/sugar/free/sightremote/database/BolusDelivered.java
@@ -16,10 +16,10 @@ import sugar.free.sightparser.applayer.descriptors.HistoryBolusType;
 public class BolusDelivered {
 
     @DatabaseField(generatedId = true)
-    private int id;
+    private long id;
 
     @DatabaseField
-    private int eventNumber;
+    private long eventNumber;
 
     @DatabaseField
     private String pump;
