+++ app/src/main/java/sugar/free/sightremote/database/EndOfTBR.java
@@ -15,10 +15,10 @@ import lombok.Setter;
 public class EndOfTBR {
 
     @DatabaseField(generatedId = true)
-    private int id;
+    private long id;
 
     @DatabaseField
-    private int eventNumber;
+    private long eventNumber;
 
     @DatabaseField
     private String pump;
